Clean up test organization and fix GC issue
