+++ tests/plotting/mappers/test_mapper.py
@@ -6,6 +6,8 @@ import pytest
 import pyvista as pv
 from pyvista.plotting import _vtk
 from pyvista.plotting.mapper import DataSetMapper
+from pyvista.plotting.utilities import algorithms
+from tests.plotting.conftest import get_actor_mapper_input
 
 
 @pytest.fixture
@@ -394,3 +396,217 @@ def test_as_rgba_uses_mapped_scalars(sphere):
     # Calling as_rgba again should be a no-op (already direct)
     mapper.as_rgba()
     assert mapper.color_mode == 'direct'
+
+
+def test_shared_mesh_raw_numpy_scalars_smooth_shading_subplots_mapper_output():
+    """Shared meshes keep distinct raw NumPy scalar arrays per subplot."""
+    n_row, n_col = 2, 2
+    pl = pv.Plotter(shape=(n_row, n_col))
+    mesh = pv.Sphere(radius=1.0, theta_resolution=80, phi_resolution=80)
+    original_point_arrays = set(mesh.point_data.keys())
+
+    actors = []
+    expected = []
+    for ii in range(n_row * n_col):
+        row, col = divmod(ii, n_col)
+        pl.subplot(row, col)
+        data = np.linalg.norm(mesh.points, axis=1) if ii == 3 else mesh.points[:, ii]
+        expected.append(np.asarray(data).copy())
+        actors.append(
+            pl.add_mesh(
+                mesh,
+                scalars=data,
+                smooth_shading=True,
+                show_scalar_bar=False,
+                ambient=0.2,
+                specular=0.3,
+                n_colors=9,
+                rng=(-1, 1.0),
+            )
+        )
+
+    names = []
+    for actor, array in zip(actors, expected, strict=True):
+        mapped = get_actor_mapper_input(actor)
+        names.append(mapped.point_data.active_scalars_name)
+        assert mapped.point_data.active_scalars_name is not None
+        np.testing.assert_allclose(
+            mapped.point_data[mapped.point_data.active_scalars_name],
+            array,
+        )
+
+    assert len(set(names)) == len(names)
+    assert mesh.point_data.active_scalars_name is None
+    assert 'Normals' in original_point_arrays or 'Normals' not in mesh.point_data
+    assert set(mesh.point_data.keys()) == original_point_arrays.union(names)
+
+    pl.close()
+
+
+def test_add_mesh_smooth_shading_with_algorithm_and_scalars_mapper_output():
+    mesh = pv.Wavelet()
+    mesh.point_data['z'] = mesh.points[:, 2].astype(float)
+    mesh.point_data['keep_active'] = mesh.points[:, 0].astype(float)
+    mesh.set_active_scalars('keep_active')
+
+    source = algorithms.source_algorithm(lambda: mesh, output_type=type(mesh))
+    surface = algorithms.callback_algorithm(
+        source,
+        lambda m: m.clip('x'),
+        output_type=pv.UnstructuredGrid,
+    )
+
+    pl = pv.Plotter()
+    actor = pl.add_mesh(
+        surface, scalars='z', smooth_shading=True, split_sharp_edges=True, show_scalar_bar=False
+    )
+    mapped = get_actor_mapper_input(actor)
+    assert mapped.point_data.active_scalars_name == 'z'
+    assert mapped.point_data.active_normals_name == 'Normals'
+    assert mapped.point_data.active_normals is not None
+    assert mesh.active_scalars_name == 'keep_active'
+    assert 'Normals' not in mesh.point_data
+
+    mesh.point_data['z'][:] = 0
+    source.Modified()
+    mapped_after = get_actor_mapper_input(actor)
+    assert np.allclose(mapped_after.point_data['z'], 0.0)
+    assert mapped_after.point_data.active_scalars_name == 'z'
+    assert mapped_after.point_data.active_normals_name == 'Normals'
+
+    pl.close()
+
+
+def test_add_mesh_smooth_shading_with_algorithm_and_scalars_propagates_updates():
+    mesh = pv.Wavelet()
+    mesh.point_data['z'] = mesh.points[:, 2].astype(float)
+
+    source = algorithms.source_algorithm(lambda: mesh, output_type=type(mesh))
+    surface = algorithms.callback_algorithm(
+        source,
+        lambda m: m.clip('x'),
+        output_type=pv.UnstructuredGrid,
+    )
+
+    pl = pv.Plotter()
+    actor = pl.add_mesh(
+        surface, scalars='z', smooth_shading=True, split_sharp_edges=True, show_scalar_bar=False
+    )
+
+    mesh.point_data['z'][:] = 0
+    source.Modified()
+
+    mapped = get_actor_mapper_input(actor)
+    assert np.allclose(mapped.point_data['z'], 0.0)
+    assert mapped.point_data.active_scalars_name == 'z'
+    assert mapped.point_data.active_normals_name == 'Normals'
+
+    pl.close()
+
+
+def test_add_mesh_smooth_shading_algorithm_raw_numpy_scalars_mapper_output():
+    cone = pv.Cone(resolution=90, capping=True).triangulate()
+    cone.clear_data()
+    cone.point_data['keep_active'] = np.linspace(-1.0, 1.0, cone.n_points, dtype=np.float32)
+    cone.set_active_scalars('keep_active')
+    source = algorithms.source_algorithm(lambda: cone, output_type=pv.PolyData)
+    raw_scalars = np.linspace(0.0, 1.0, cone.n_points, dtype=np.float32)
+
+    pl = pv.Plotter()
+    actor = pl.add_mesh(
+        source,
+        scalars=raw_scalars,
+        smooth_shading=True,
+        split_sharp_edges=True,
+        show_scalar_bar=False,
+    )
+    mapped = get_actor_mapper_input(actor)
+    tracker = np.asarray(
+        mapped.point_data[algorithms.SmoothShadingAlgorithm.ORIGINAL_POINT_IDS_NAME]
+    )
+    assert mapped.point_data.active_scalars_name == pv.DEFAULT_SCALARS_NAME
+    assert mapped.point_data.active_normals_name == 'Normals'
+    assert np.allclose(mapped.point_data[pv.DEFAULT_SCALARS_NAME], raw_scalars[tracker])
+    assert cone.active_scalars_name == 'keep_active'
+    assert pv.DEFAULT_SCALARS_NAME not in cone.point_data
+    assert 'Normals' not in cone.point_data
+
+    pl.close()
+
+
+def test_add_mesh_raw_numpy_cell_scalars_mapper_output():
+    sphere = pv.Sphere()
+    sphere.cell_data['keep_active'] = -np.arange(sphere.n_cells, dtype=np.float32)
+    sphere.set_active_scalars('keep_active', preference='cell')
+    original_active_normals_name = sphere.point_data.active_normals_name
+    cell_scalars = np.arange(sphere.n_cells, dtype=np.float32)
+
+    pl = pv.Plotter()
+    actor = pl.add_mesh(sphere, scalars=cell_scalars, show_scalar_bar=False)
+    mapped = get_actor_mapper_input(actor)
+    assert mapped.cell_data.active_scalars_name == pv.DEFAULT_SCALARS_NAME
+    assert np.allclose(mapped.cell_data[pv.DEFAULT_SCALARS_NAME], cell_scalars)
+    assert sphere.cell_data.active_scalars_name == 'keep_active'
+    assert sphere.point_data.active_normals_name == original_active_normals_name
+
+    pl.close()
+
+
+def test_add_mesh_smooth_shading_multi_component_cell_scalars_mapper_output():
+    sphere = pv.Sphere()
+    cell_vec = np.column_stack(
+        [
+            np.zeros(sphere.n_cells, dtype=np.float32),
+            np.linspace(-1.0, 1.0, sphere.n_cells, dtype=np.float32),
+            np.zeros(sphere.n_cells, dtype=np.float32),
+        ]
+    )
+    sphere.cell_data['keep_active'] = np.arange(sphere.n_cells, dtype=np.float32)
+    sphere.set_active_scalars('keep_active', preference='cell')
+    original_active_normals_name = sphere.point_data.active_normals_name
+
+    pl = pv.Plotter()
+    actor = pl.add_mesh(
+        sphere,
+        scalars=cell_vec,
+        component=1,
+        smooth_shading=True,
+        show_scalar_bar=False,
+    )
+    mapped = get_actor_mapper_input(actor)
+    derived_name = f'{pv.DEFAULT_SCALARS_NAME}-1'
+    assert mapped.cell_data.active_scalars_name == derived_name
+    assert mapped.point_data.active_normals_name == 'Normals'
+    assert np.allclose(mapped.cell_data[derived_name], cell_vec[:, 1])
+    assert sphere.cell_data.active_scalars_name == 'keep_active'
+    assert sphere.point_data.active_normals_name == original_active_normals_name
+
+    pl.close()
+
+
+@pytest.mark.skip_check_gc  # Mapper-array inspection retains a vtkWeakReference.
+def test_add_mesh_smooth_shading_multi_component_scalars_mapper_output():
+    mesh = pv.Sphere(theta_resolution=60, phi_resolution=60).cast_to_unstructured_grid()
+    mesh.clear_data()
+    mesh.point_data['vec'] = np.array(mesh.points, dtype=np.float32)
+    mesh.point_data['keep_active'] = mesh.points[:, 0].astype(np.float32)
+    mesh.set_active_scalars('keep_active')
+
+    pl = pv.Plotter()
+    actor = pl.add_mesh(
+        mesh, scalars='vec', component=1, smooth_shading=True, show_scalar_bar=False
+    )
+    actor.mapper.update()
+    mapped = actor.mapper.GetInputDataObject(0, 0)
+    point_data = mapped.GetPointData()
+    assert point_data.GetScalars().GetName() == 'vec-1'
+    assert point_data.GetNormals().GetName() == 'Normals'
+    assert np.allclose(
+        _vtk.vtk_to_numpy(point_data.GetArray('vec-1')),
+        _vtk.vtk_to_numpy(point_data.GetArray('vec'))[:, 1],
+    )
+    assert mesh.active_scalars_name == 'keep_active'
+    assert 'Normals' not in mesh.point_data
+
+    pl.close()
+    del point_data, mapped, actor, pl
